requirements.txt
# Llama Stack and AI
llama-stack-client>=0.0.50
llama-stack>=0.0.50

# Vector Store and Embeddings
chromadb>=0.4.22
sentence-transformers>=2.3.1

# HTTP and API
httpx>=0.26.0
requests>=2.31.0
aiohttp>=3.9.0

# Configuration and Environment
pyyaml>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# MCP Integration
mcp>=0.1.0

# Utilities
python-dateutil>=2.8.2
jinja2>=3.1.2
markdown>=3.5.0

# Logging and Monitoring
structlog>=23.3.0
prometheus-client>=0.19.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=23.12.0
ruff>=0.1.9
mypy>=1.8.0
pre-commit>=3.6.0

# Optional: Web Interface
fastapi>=0.109.0
uvicorn[standard]>=0.26.0
